Refactored FactorySpecification into FactoryTestHelper according to @johanneshiry's suggestion
