+++ src/test/groovy/edu/ie3/test/helper/FactorySpecification.groovy
@@ -1,11 +1,10 @@
-package edu.ie3.io.factory
+package edu.ie3.test.helper
 
-import spock.lang.Specification
 import tec.uom.se.quantity.Quantities
 
 import javax.measure.Unit
 
-class FactorySpecification extends Specification {
+trait FactoryTestHelper {
     static getQuant(String parameter, Unit unit) {
         return Quantities.getQuantity(Double.parseDouble(parameter), unit)
     }
